CMake Nade and Sandbox separated

+++ Nade/src/Graphics/Components/Renderer.h
@@ -10,7 +10,7 @@ namespace Nade {
 	public:
 		void Draw(GameObject& object, Shader* shader) {
 			if (object.is_physics_object) {
-				object.velocity += ND_PHYSICS_GRAVITATIONAL_ACCELERATION;
+				object.velocity += (float)ND_PHYSICS_GRAVITATIONAL_ACCELERATION;
 			}
 			object.Translate(0, object.velocity, 0);
 			RawModel model = object.GetData().model;
